+++ CHANGELOG.md
@@ -4,9 +4,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+- indexHtmlPath config setting
 
-## [0.5.0-rc1] 
+## [0.5.0-rc1-9] 
 
+- normalization of fail/succes/request/end actions 
 - change: module interfaces, everything is provided through main index.js of module
 - new: ufp-core cli command main command with stats,build,start,status commands for now managing ufp app 
     try it out with 'ufp-core' or 'node_modules/.bin/ufp-core' in case your binaries are not linked,